requirements.txt
streamlit
vosk
pyaudio
requests
chromadb
sentence-transformers
PyMuPDF
langchain-community
numpy
pandas
faster-whisper
rank_bm25
langchain-text-splitters
google-genai
python-dotenv
edge-tts
mcp
